refactor(cli): split render layer out of main.rs
First incremental step toward decomposing the 6.3k-line main.rs
(the long-standing SRP-major finding). Moves the rendering layer
into a new render/ submodule:

- render/graph.rs (1,179 lines): Cytoscape JSON builders and their
  support types - CytoscapeElement, CytoscapeData, CytoscapePosition,
  StateElementParams; functions collect_action_info, classify_action,
  build_automaton_compound_node, build_state_elements,
  counterstrategy_to_cytoscape, dsl_automata_to_cytoscape,
  unrolled_automata_to_cytoscape, generate_cytoscape_html;
  Cytoscape-private helpers strip_outer_parens, extract_literal_value,
  expr_to_string, expr_to_string_inner.

- render/text.rs (223 lines): human-readable formatters -
  render_proposal_provenance, render_discharge_verdict_text (+ its
  indented variant), render_controller_diagnostics.

- render/mod.rs declares the two submodules.

main.rs: 6,324 -> 4,940 lines (-1,384). Types and functions used
by main.rs are pub(crate); items only used inside graph.rs / text.rs
stay private to their submodule. No external consumers exist
(mununu-cli is a binary crate), so nothing needs to escape pub(crate).

Behavior unchanged: Cytoscape JSON and text rendering produce
byte-identical output (same caller sites, same arguments). All 1,672
tests pass; clippy -D warnings clean.

Deferred for follow-up sessions: loader.rs (load_with_adapter_mode,
load_context_documents_mode), cmd/*.rs (per-subcommand handlers).
