+++ crates/mununu-cli/src/render/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod graph;
+pub(crate) mod text;
